Move table column utility
